lazy tests: aggregation queries to separate module
